Update simplewall.lng
Change "paquets abandonnées" to "paquet bloqués" 
Cause it's confusing when look for stop notification. "Paquet abandonnées" is never used in any fire-wall software. 
In french Packet are blocked not dropped ;-)
Plus add space after question mark (in french is like that)
file was UTF-16LE and github seems want to convert it to UTF-8 so be carefull