Update French.ini
Hello!

Here are the following changes in French:
- correcting the syntax for "!", "?" and ":" as they require a space before
- removing capital letters in the middle of the phrase (never in French unless it is a name but not to talk about things)
- adding the missing translations
- correcting a few grammatical errors
- adjusted some text to reflect what is usually displayed in comparable apps (Remerciements! is Dire merci ! for instance)
- adjusted some elements to harmonize globally (example: "règle de l'utilisateur" or "règle d'utilisateur")

All changes are based on https://github.com/henrypp/simplewall/blob/master/bin/i18n/!example.txt that I used as the English reference (I hope it's the right file for that).

There is one element I can't put my finger on: IDS_LAYER
I can't find in the app where it is. I can't translate it without a bit of context so I don't know for this one. 

In addition, there are a few other mistakes in SimpleWall that can't be corrected here as they combine several of these elements which don't always work correctly in French. 
Maybe this should be dealt on the side. I'll be happy to help if needed. 

Thanks!
Have a good one.